+++ example/json.dart
@@ -4,6 +4,7 @@ import 'dart:isolate';
 
 import 'package:angel_container/mirrors.dart';
 import 'package:angel_framework/angel_framework.dart';
+import 'package:angel_framework/http.dart';
 
 main() async {
   int x = 0;