+++ example/view.dart
@@ -1,5 +1,6 @@
 import 'package:angel_container/mirrors.dart';
 import 'package:angel_framework/angel_framework.dart';
+import 'package:angel_framework/http.dart';
 
 main() async {
   var app = new Angel(reflector: MirrorsReflector());