+++ lib/angel_framework.dart
@@ -1,5 +1,8 @@
 /// An easily-extensible web server framework in Dart.
 library angel_framework;
 
+export 'package:angel_http_exception/angel_http_exception.dart';
+export 'package:angel_model/angel_model.dart';
+export 'package:angel_route/angel_route.dart';
+export 'package:body_parser/body_parser.dart' show FileUploadInfo;
 export 'src/core/core.dart';
-export 'src/http/http.dart';