+++ test/controller_test.dart
@@ -4,6 +4,7 @@ import 'dart:io';
 
 import 'package:angel_container/mirrors.dart';
 import 'package:angel_framework/angel_framework.dart';
+import 'package:angel_framework/http.dart';
 import 'package:http/http.dart' as http;
 import 'package:mock_request/mock_request.dart';
 import 'package:test/test.dart';