+++ test/detach_test.dart
@@ -1,5 +1,6 @@
 import 'dart:convert';
 import 'package:angel_framework/angel_framework.dart';
+import 'package:angel_framework/http.dart';
 import 'package:mock_request/mock_request.dart';
 import 'package:test/test.dart';