+++ test/general_test.dart
@@ -1,6 +1,7 @@
 import 'dart:io';
 import 'package:angel_container/mirrors.dart';
 import 'package:angel_framework/angel_framework.dart';
+import 'package:angel_framework/http.dart';
 import 'dart:convert';
 import 'package:http/http.dart' as http;
 import 'package:test/test.dart';